requirements.txt
python-telegram-bot
yfinance
python-dotenv
